+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+* Added react client.
+* Added issue log.
+* Removed eslint-vue-config dev dependency.
+* updated git ignore to include dev folders.
+
+## 0.1.11
+
 * Added changelog file.
 * Added beginnings of attributions.
 * Added license and repository to package.json.